pyproject.toml
[project]
name = "evergreen-tvevents"
version = "1.0.0"
description = "High-throughput TV telemetry ingestion microservice."
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "psycopg2-binary==2.9.10",
    "jsonschema==3.2.0",
    "boto3==1.38.14",
    "botocore==1.38.14",
    "requests>=2.32.4",
    "PyYAML==6.0.2",

    "opentelemetry-api==1.31.1",
    "opentelemetry-distro==0.52b1",
    "opentelemetry-exporter-otlp-proto-http",
    "opentelemetry-instrumentation==0.52b1",
    "opentelemetry-instrumentation-fastapi==0.52b1",
    "opentelemetry-instrumentation-psycopg2==0.52b1",
    "opentelemetry-instrumentation-requests==0.52b1",
    "opentelemetry-instrumentation-logging==0.52b1",
    "opentelemetry-instrumentation-urllib==0.52b1",
    "opentelemetry-instrumentation-urllib3==0.52b1",
    "opentelemetry-instrumentation-asgi==0.52b1",
    "opentelemetry-instrumentation-threading==0.52b1",
    "opentelemetry-instrumentation-dbapi==0.52b1",
    "opentelemetry-propagator-aws-xray==1.0.2",
    "opentelemetry-sdk==1.31.1",
    "opentelemetry-semantic-conventions==0.52b1",
    "opentelemetry-util-http==0.52b1",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.5",
    "pytest-cov==6.2.1",
    "pytest-asyncio==0.24.0",
    "pytest-env",
    "httpx>=0.27.0",
    "coverage[toml]==7.10.1",
    "ruff>=0.8.0",
    "mypy",
    "types-jsonschema",
    "types-psycopg2",
    "types-requests",
    "complexipy>=0.4.0",
    "pip-tools==7.5.0",
    "pip==24.3.1",
    "setuptools==78.1.1",
]

[build-system]
requires = ["setuptools>=78"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"app" = "src/app"}
py-modules = ["app"]

[tool.setuptools.packages.find]
where = ["src"]
namespaces = false

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "S", "B", "A", "C4", "SIM"]
ignore = ["S101", "S104"]

[tool.mypy]
ignore_missing_imports = true
